requirements.txt
gradio
insightface==0.7.3
numpy==1.26.4
opencv-python
Pillow==9.0.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
onnxruntime-gpu==1.15.1
scikit-image==0.19.3
matplotlib==3.5.1
scikit-learn
fastapi